Redesign of the profile page
