--- main/templates/profile.html
@@ -1,59 +1,0 @@
-# extends 'base.html'
-# import 'macro/forms.html' as forms
-# import 'macro/utils.html' as utils
-
-# block content
-  <div class="page-header">
-    <h1>
-      <small class="pull-right">{{utils.auth_icons(user_db)}}</small>
-      {{title}}
-    </h1>
-  </div>
-  <form method="POST" action="." novalidate>
-    <div class="row">
-      <div class="col-md-4 col-sm-6">
-        {{form.csrf_token}}
-        <div class="form-group">
-          <label class="control-label" for="username">Username</label>
-          <input id="username" class="form-control" name="username" type="text" value="{{user_db.username}}" readonly="readonly">
-        </div>
-        {{forms.text_field(form.name, autofocus=True)}}
-        # include 'user/user_email_field.html'
-
-        # if config.CONFIG_DB.has_email_authentication
-          # if user_db.password_hash
-            <input class="chrome-autocomplete-hack hide">
-            {{forms.password_visible_field(form.old_password)}}
-            # if form.old_password.errors
-              <div class="alert alert-info">
-                <a class="alert-link" href="{{url_for('user_forgot')}}" target="_blank">Forgot your password?</a>
-              </div>
-            # endif
-          # endif
-          {{forms.password_visible_field(form.new_password)}}
-        # endif
-      </div>
-      <div class="col-md-2 col-sm-3">
-        <div class="form-group">
-          <label class="control-label">Avatar</label>
-          <div>
-            <img class="img-thumbnail" src="{{user_db.avatar_url_size(170)}}" alt="Avatar of {{user_db.name}}">
-          </div>
-          <p class="help-block">
-            Change on
-            <a href="//gravatar.com" target="_blank">Gravatar</a>
-          </p>
-        </div>
-      </div>
-    </div>
-    <div class="row">
-      <div class="col-md-6 col-sm-9">
-        <div class="controls">
-          <button type="submit" class="btn btn-primary btn-block btn-loading" data-loading-text="Updating Profile..">
-            Update Profile
-          </button>
-        </div>
-      </div>
-    </div>
-  </form>
-# endblock
